requirements.txt
# Core Data Science
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.6.1

# Machine Learning
scikit-learn>=1.4.0
pycaret>=3.3.0
imbalanced-learn>=0.12.0
category-encoders>=2.4.0
lightgbm>=3.0.0
pyod>=1.1.3

# Experiment Tracking
mlflow<2.0

# AI/LLM Integration
openai>=1.0.0
python-dotenv>=1.0.0

# Data Processing
tqdm>=4.62.0
joblib>=1.2.0

# Jupyter & Visualization
jupyter>=1.0.0
ipython>=5.5.0
ipywidgets>=7.6.5
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
click>=7.0
requests>=2.17.3
pyyaml>=5.1
pytz>=2022.7.1
packaging>=21.3
sqlparse>=0.4.0
alembic<2
protobuf>=3.12.0
gitpython>=2.1.0
entrypoints>=0.4
cloudpickle<3
databricks-cli>=0.8.7
